Start to extract watch to its own module
